requirements.txt
asgiref==3.11.1
beautifulsoup4==4.14.3
colorama==0.4.6
Django==6.0.3
django-bootstrap4==26.1
django-countries==8.2.0
django-model-utils==5.0.0
django-phonenumber-field==8.4.0
phonenumbers==9.0.27
pillow==12.1.1
python-dotenv==1.2.2
six==1.17.0
soupsieve==2.8.3
sqlparse==0.5.5
swapper==1.4.0
tqdm==4.67.3
typing_extensions==4.15.0
tzdata==2025.3
